test: delete the storefront and route-URL cases another file already owns
url.test.ts re-tested extractStorefrontFromURL and three URL builders
that storefront.test.ts covers, leaving only its second describe doing
what its header claims. itms.test.ts table-tested buildItmsRouteURL
against the same five mappings storefront.test.ts asserts inside an
active-service loop that strictly contains them. storefront.test.ts kept
its count at 41, so nothing moved out of coverage.
